Split Sobel into header and source file
